Add JSONSerializable and default option
JSONSerializable class provide abstract method `from_dict` and
`to_dict`, required by JSONDeserializer and JSONSerializer to
serialize object to json bytes and vice versa.
Inheriting the JSONSerializable class, these two method need to
be defined by the subclass in order to use JSON serialization.

Serializable is renamed to SerializationMixin, to remove confusion
about naming and keep it serparate from JSONSerializable.

Option for default serializer and deserializer are added to
SerializatioMixin. The method default_serializer will be defined
to provide a default serializer for serialization, if a serializer
is not provided (similar for default_deserializer).

By using this method, Metadata or other container can define a
default option in case of multiple serializer (or deserialzer)
other than JSON one, which makes it simpler. Otherwise, have to
redefine all methods of Mixin for each Serialization method.

Signed-off-by: Pradyumna Krishna <git@onpy.in>