Add match_payload method to Envelope
match_payload method checks for the payoad if deserializes into
the provided class type. This method is essential in order to
add DSSE compatibility in in_toto.

This method will provide interface similar to current in_toto's
Metablock property `_type`. Which is used at various places in
in_toto project to identify type of the payload.

Note: This type is not payload_type of DSSE envelope.

Signed-off-by: Pradyumna Krishna <git@onpy.in>